refactor: decompose card_image_display/handlers.py into focused mixins (#549)
Split the 385-LOC CardImageDisplayHandlersMixin into single-concern peer
mixins composed in frame.py, and add a protocol.py typing the shared self
surface so each new mixin inherits a TYPE_CHECKING _Base instead of
re-declaring shared attributes:

- image_loader.py  -> _ImageLoaderMixin  (off-thread decode/scale/mask pipeline)
- animation.py     -> _AnimationMixin    (fade-transition timer loop)
- bitmap_renderer.py -> _BitmapRendererMixin (canvas/border/flip-icon/placeholder)
- protocol.py      -> CardImageDisplayProto (cross-mixin self contract)

handlers.py keeps only the public state API (show_*/flip-icon) and the
input/event callbacks, dropping from 385 to ~120 LOC. No behavior change:
method bodies were moved verbatim.
